% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_aspe.R
\name{clean_aspe}
\alias{clean_aspe}
\title{Mise d'un df brut aspe au format passerelle}
\usage{
clean_aspe(df)
}
\arguments{
\item{df}{Dataframe contenant les données.}
}
\value{
Un dataframe au format souhaité avec les coordonnées reprojetées
}
\description{
Mise d'un df brut aspe au format passerelle
}
\examples{
\dontrun{
aspe_propre <- aspe_brut \%>\%
clean_aspe()
}
}
