% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fede.R
\name{clean_fede}
\alias{clean_fede}
\title{Mise d'un df brut des pêches fédés au format passerelle}
\usage{
clean_fede(df_brut, crs_init = 2154, crs_fin = 4326)
}
\arguments{
\item{df_brut}{Dataframe données SD.}

\item{crs_init}{Numérique. Code EPSG du CRS initial. Par défaut c'est 2154 (Lambert 93).}

\item{crs_fin}{Numérique. Code EPSG du CRS de sortie. Par défaut c'est 4326 (WGS84).}
}
\value{
Un dataframe au format souhaité avec les coordonnées reprojetées
}
\description{
Mise d'un df brut des pêches fédés au format passerelle
}
\examples{
\dontrun{
sd_fede_propre <- sd_fede_brut \%>\%
clean_fede()
}
}
