% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fede22.R
\name{clean_fede22}
\alias{clean_fede22}
\title{Fonction permettant de nettoyer le dataframe brut obtenu avec la fonction
lire_xls_fede22() en l'homogénéisant par rapport aux autres dataframes.}
\usage{
clean_fede22(df_brut, crs_init = 2154, crs_fin = 4326)
}
\arguments{
\item{df_brut}{Caractère. Fichier désiré à nettoyer.}

\item{crs_init}{Ne pas renseigner. Crs initial du fichier, déjà renseigné.}

\item{crs_fin}{Ne pas renseigner. Crs final du fichier, déjà renseigné.}
}
\value{
Un dataframe homogénéisé avec les autres dataframes des autres données
de pêche
}
\description{
Fonction permettant de nettoyer le dataframe brut obtenu avec la fonction
lire_xls_fede22() en l'homogénéisant par rapport aux autres dataframes.
}
\examples{
\dontrun{
fede_propre <- fede_brut \%>\%
clean_fede22()
}
}
