% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sd.R
\name{clean_sd}
\alias{clean_sd}
\title{Mise d'un df brut des pêches SD au format passerelle}
\usage{
clean_sd(sd_brut, crs_init = 2154, crs_fin = 4326)
}
\arguments{
\item{sd_brut}{Dataframe données SD}

\item{crs_init}{Numérique. Code EPSG du CRS initial}

\item{crs_fin}{Numérique. Code EPSG du CRS de sortie}
}
\value{
Un dataframe au format souhaité avec les coordonnées reprojetées
}
\description{
Mise d'un df brut des pêches SD au format passerelle
}
\examples{
\dontrun{
sd_data__propre <- sd_data_brut \%>\%
clean_sd()
}
}
