% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wama.R
\name{clean_wama}
\alias{clean_wama}
\title{Mise d'un df wama brut au format passerelle}
\usage{
clean_wama(wama_brut, crs_init = 2154, crs_fin = 4326)
}
\arguments{
\item{wama_brut}{Dataframe wama}

\item{crs_init}{Numérique. Code EPSG du CRS initial}

\item{crs_fin}{Numérique. Code EPSG du CRS de sortie}
}
\value{
Un dataframe au format souhaité avec les coordonnées reprojetées
}
\description{
Mise d'un df wama brut au format passerelle
}
\examples{
\dontrun{
wama_propre <- wama_brut \%>\%
clean_wama()
}
}
