% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donner_statut_sp_bv.R
\name{donner_statut_sp_bv}
\alias{donner_statut_sp_bv}
\title{Donner le statut des especes présence/absence dans chaque bassin versant, indiquant en plus si
l'absence est avérée ou est due à un échantillonnage insuffisant.}
\usage{
donner_statut_sp_bv(sf_data)
}
\arguments{
\item{sf_data}{dataframe de données sf}
}
\value{
un dataframe au format souhaité avec le code station, le code espèce et le statut
}
\description{
Donner le statut des especes présence/absence dans chaque bassin versant, indiquant en plus si
l'absence est avérée ou est due à un échantillonnage insuffisant.
}
\examples{
statut_bv <- donner_statut_sp_bv(sf_data = data)
}
