% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donner_statut_sp_point.R
\name{donner_statut_sp_point}
\alias{donner_statut_sp_point}
\title{Donner le statut des especes présence/absence dans chaque station}
\usage{
donner_statut_sp_point(sf_data)
}
\arguments{
\item{sf_data}{dataframe de données sf}
}
\value{
un dataframe au format souhaité avec le code station, le code espèce
les effectifs, le statut et le type de pêche
}
\description{
Donner le statut des especes présence/absence dans chaque station
}
\examples{
\dontrun{
statut <- donner_statut_sp_point(sf_data = data)
}
}
