% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expliciter_statut_lr.R
\name{expliciter_statut_lr}
\alias{expliciter_statut_lr}
\title{Fonction permettant d'expliciter le statut liste rouge de chaque espèce}
\usage{
expliciter_statut_lr(df, var)
}
\arguments{
\item{df}{le dataframe}

\item{var}{la variable à expliciter}
}
\value{
modification du dataframe avec modification des valeurs de la variable
}
\description{
Fonction permettant d'expliciter le statut liste rouge de chaque espèce
}
\examples{
expliciter_statut_lr(df = passerelle_taxo, var = lr_nationale)

}
