% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coords.R
\name{get_coords}
\alias{get_coords}
\title{Collecte et reprojection des coordonnées}
\usage{
get_coords(
  sf_obj,
  crs_init,
  crs_fin = 4326,
  col_names = c("x_wgs84", "y_wgs84")
)
}
\arguments{
\item{sf_obj}{Objet de classe sf dont il s'agit de collecter les coordonnées.}

\item{crs_init}{Numérique. Code EPSG du système de coordonnées de cet objet.}

\item{crs_fin}{Numérique. Code EPSG du système de coordonnées de sortie. Par défaut, 4326
pour une sortie en WGS84.}

\item{col_names}{Vecteur caractères de deux éléments pour nommer les colonnes de coordonnées.
Par défaut elles sont nommées "x_wgs84" et "y_wgs84".}
}
\value{
Un dataframe de deux colonnes avec les coordonnées dans le CRS de sortie.
}
\description{
Collecte et reprojection des coordonnées
}
\examples{
\dontrun{
# Les coordonnees de la DR
x <- 304660
y <- 2353600

# Construction de l'objet sf correspondant
dr_ofb <- sf::st_sfc(sf::st_point(c(x,y)))
sf::st_crs(dr_ofb) <- 27572
dr_ofb

# Coordonnees de la DR en WGS84
get_coords(dr_ofb, crs_init = 27572)
}
}
