% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_fichier_fedes.R
\name{lire_fichier_fedes}
\alias{lire_fichier_fedes}
\title{Fonction permettant la lecture des données des fédés départementales}
\usage{
lire_fichier_fedes(chemin, depts = c("22", "35", "29", "56"))
}
\arguments{
\item{chemin}{Caractère. Chemin vers le fichier de xonnées}

\item{depts}{Vecteur caractères. Indique les départements à premdre en compte,
c'est-à-dire les noms des onglets du fichier de données.}
}
\value{
Le dataframe
}
\description{
Fonction permettant la lecture des données des fédés départementales
}
\examples{
\dontrun{
fedes <- lire_fichier_fedes(chemin = "raw_data/fedes_departementales_peche.xlsx")
}
}
