% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_xls_fede22.R
\name{lire_xls_fede22}
\alias{lire_xls_fede22}
\title{Fonction permettant la lecture des données de la fédé22.}
\usage{
lire_xls_fede22(fichier)
}
\arguments{
\item{fichier}{charactère. Donner ici le fichier à lire}
}
\value{
Un dataframe à partir du fichier xls
}
\description{
Fonction permettant la lecture des données de la fédé22.
}
\examples{
\dontrun{
data_f22 <- lire_xls_fede22("raw_data/donnees_fede22/FDPPMA_22_peches_scientifiques_2021.xls")
}
}
