% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_xlsx_fede35.R
\name{lire_xlsx_fede35}
\alias{lire_xlsx_fede35}
\title{Assemblage des fichiers annuels}
\usage{
lire_xlsx_fede35(
  repertoire,
  fichier_reference = "CR op pêche elec FD35 2020-VF.xlsx"
)
}
\arguments{
\item{repertoire}{Caractère. Chemin vers le répertoire où se trouvent les fichiers Excel de données.}

\item{fichier_reference}{Caractère. Nom du fichier excel dont les noms de colonnes vont servir
de référence pour homogénéiser les noms des colonnes des différents fichiers excel}
}
\value{
Dataframe pour l'ensemble des années de données disponibles par la fédé 35.
}
\description{
Cette fonction permet d'assembler - en les empilant - les fichiers xlsx en homogénéisant
    les noms des colonnes en un unique dataframe.
}
\examples{
\dontrun{
data <- lire_xlsx_fede35(repertoire="raw_data",
fichier_reference = "CR op pêche elec FD35 2020-VF.xlsx")
)
}

}
