% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_and_filter_species.R
\name{recode_and_filter_species}
\alias{recode_and_filter_species}
\title{Recodage des codes espèces et exclusion de taxons}
\usage{
recode_and_filter_species(df, sp_to_remove = NA)
}
\arguments{
\item{df}{Dataframe standardisé contenant les données.}

\item{sp_to_remove}{Vecteur texte contenant les codes à trois lettres des taxons à
supprimer.}
}
\value{
Le dataframe mis à jour.
}
\description{
Par exemple les carpes cuir, miroir, etc. sont regroupées sous un unique code CCX.
Idem pour les vandoises en VAX.
Dans l'ouest Finistère, il n'y a que de l'épinoche => recodage de l'épinochette sur cette zone.
Tous les carassins sont des argentés.
}
\examples{
\dontrun{
df_propre <- df_brut \%>\%
recode_and_filter_species(sp_to_remove = c("OCV", "ASA"))
}
}
