% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_crs.R
\name{transform_crs}
\alias{transform_crs}
\title{Reprojeter un df aspe}
\usage{
transform_crs(
  aspe_df,
  coords = c("pop_coordonnees_x", "pop_coordonnees_y"),
  crs_init,
  crs_fin = 4326,
  coord_names = c("x_wgs84", "y_wgs84")
)
}
\arguments{
\item{aspe_df}{Dataframe Aspe issu de clean_aspe().}

\item{coords}{Vecteur contenant les noms des colonnes de coordonnées, X puis Y dans aspe_df.}

\item{crs_init}{Numérique. Code EPSG du CRS initial.}

\item{crs_fin}{Numérique. Code EPSG du CRS de sortie. Par défaut c'est 4326 (WGS84).}

\item{coord_names}{Vecteur contenant les noms des colonnes de coordonnées, X puis Y en sortie.}
}
\value{
Le dataframe avec les coordonnées reprojetées en colonnes
}
\description{
Reprojeter un df aspe
}
\examples{
\dontrun{
aspe_l2 <- aspe_occurence \%>\%
filter(proj_pop == "Lambert II Etendu") \%>\%
transform_crs(crs_init = 27572, crs_fin = 4326)
}
}
