% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_absence.R
\name{ajouter_absence}
\alias{ajouter_absence}
\title{Fonction permettant d'ajouter les absence des espèces pour chaque opération}
\usage{
ajouter_absence(df)
}
\arguments{
\item{df}{fichier de donnees qui doit contenir une variable "code_espece" mais pas les
noms latin ou commun, ainsi que la variable "effectif".}
}
\value{
Le dataframe df complété avec les combinaisons manquantes entre les codes espèces et
    les autres variables.
}
\description{
Fonction permettant d'ajouter les absence des espèces pour chaque opération
}
\examples{
ajouter_absence(data)
}
