% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wama.R
\name{clean_wama}
\alias{clean_wama}
\title{Mise en forme d'un df aspe au format d'échange commun avec aspe, fédés, etc.}
\usage{
clean_wama(df_brut, crs_init = 2154, crs_fin = 4326)
}
\arguments{
\item{df_brut}{Dataframe wama.}

\item{crs_init}{Numérique. Code EPSG du CRS initial. Par défaut c'est 2154 (Lambert 93).}

\item{crs_fin}{Numérique. Code EPSG du CRS de sortie. Par défaut c'est 4326 (WGS84).}
}
\value{
Un dataframe au format souhaité avec les coordonnées reprojetées.
}
\description{
Mise en forme d'un df aspe au format d'échange commun avec aspe, fédés, etc.
}
\examples{
\dontrun{
wama_propre <- wama_brut \%>\%
clean_wama()
}
}
