% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_xlsx_fede35.R
\name{lire_xlsx_fede35}
\alias{lire_xlsx_fede35}
\title{Assemblage des fichiers annuels}
\usage{
lire_xlsx_fede35(repertoire, fichier_reference)
}
\arguments{
\item{fichier_reference}{Caractère. Chemin vers le fichier excel dont les noms de colonnes vont servir
de référence pour homogénéiser les noms des colonnes des différents fichiers excel}
}
\value{
Dataframe national pour l'ensemble des années de données disponibles
}
\description{
Cette fonction permet d'assembler - en les empilant - les fichiers csv issus de la fonction
    telecharger_fichiers_onde_annuels() en un unique dataframe. La sortie est le dataframe national
    pour l'ensemble des années où la donnée est disponible.
}
\examples{
\dontrun{
onde <- assembler_fichiers_onde_annuels_csv(
annual_onde_files_dir = "raw_data/fichiers_onde_annuels_zippes"
)
}

}
